pyproject.toml
[project]
name = "chaddi-tg"
version = "0.1.0"
description = "Telegram bot"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.0.0",
    "youtube-dl==2021.12.17",
    "en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl",
    "cachetools>=5.3.3",
    "ciso8601>=2.3.3",
    "fastapi>=0.119.1",
    "fastapi-socketio>=0.0.10",
    "googletrans>=4.0.2",
    "ddgs>=0.1.0",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "markdown-it-py>=4.0.0",
    "markupsafe>=3.0.3",
    "nltk>=3.9.2",
    "numpy>=2.2.6",
    "openai>=2.6.0",
    "peewee>=3.18.2",
    "pillow>=12.0.0",
    "preshed>=3.0.10",
    "prometheus-client>=0.23.1",
    "psycopg2>=2.9.11",
    "pycparser>=2.23",
    "pydantic>=2.12.3",
    "pydantic-core>=2.41.4",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "python-engineio>=4.12.3",
    "python-multipart>=0.0.20",
    "python-socketio>=5.14.2",
    "python-telegram-bot[job-queue]>=22.5",
    "pytz>=2025.2",
    "pyyaml>=6.0.3",
    "rake-nltk>=1.0.6",
    "requests>=2.32.5",
    "shortuuid>=1.0.13",
    "spacy>=3.8.7",
    "spacy-legacy>=3.0.12",
    "starlette>=0.48.0",
    "starlette-exporter>=0.23.0",
    "tzlocal>=5.3.1",
    "urllib3>=2.5.0",
    "uvicorn>=0.38.0",
    "sentry-sdk[fastapi]>=2.48.0",
    "posthog>=7.4.2",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-exporter-otlp-proto-http>=1.39.1",
    "typer>=0.21.0",
    "opentelemetry-exporter-otlp>=1.39.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.14.1",
]

[tool.ruff]
# Set the target Python version
target-version = "py313"

# Exclude common directories
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

# Set line length to 100 characters
line-length = 100

[tool.ruff.lint]
# Enable these rule sets
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "RUF",    # Ruff-specific rules
]

# Ignore specific rules that might be too strict
ignore = [
    "E501",   # line too long (handled by formatter)
    "RUF006",  # asyncio.create_task reference warning
    ]

[tool.ruff.lint.per-file-ignores]
# Ignore import order issues in __init__.py files
"__init__.py" = ["F401", "F405"]
# Ignore UP017 (datetime.UTC) in version.py for production compatibility
"src/domain/version.py" = ["UP017"]

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Indent with spaces
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
